requirements.txt
parameterized==0.7.4
termcolor==1.1.0
icmplib==2.1.1
urllib3==1.25.10
netaddr==0.8.0
PyYAML==5.4.1
